Add files via upload
# Weather Data Trends

## Goal of the project
To plot trends of the temperature of a city and of the entire globe over the years, and draw similarities and differences between the same.

We draw line plots of the given data with appropriate description , labelling and legend.  

In the end we give appropriate conclusions based on our observations.